requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.6.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
chicken==0.1.0
chromedriver-binary==91.0.4472.19.0
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
discord.py==1.7.1
egg==0.2.0
environs==9.3.2
flake8==4.0.1
h11==0.12.0
idna==2.10
lxml==4.6.3
marshmallow==3.11.1
mccabe==0.6.1
multidict==5.1.0
outcome==1.1.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyOpenSSL==21.0.0
python-dotenv==0.17.0
requests==2.25.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.7.4.3
urllib3==1.26.4
webdriver-manager==3.5.2
wsproto==1.0.0
yarl==1.6.3
